requirements.txt
Flask==3.0.2
flask_cors==4.0.0
memory_profiler==0.61.0
objgraph==3.6.1
opencv_python==4.9.0.80
Pillow==10.2.0
Shapely==2.0.3
tweepy==4.14.0
ultralytics==8.3.135
absl-py==2.1.0
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.26
botocore==1.34.26
build==1.0.3
cachetools==5.3.2
cattrs==23.2.3
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
coremltools==7.1
coverage==5.3.1
cycler==0.12.1
Cython==3.0.8
Deprecated==1.2.14
distlib==0.3.8
docutils==0.17.1
einops==0.3.2
exceptiongroup==1.2.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
google-auth==2.26.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
humanfriendly==10.0
hydra-core==1.3.2
identify==2.5.33
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json-tricks==3.16.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lapx==0.5.5
loguru==0.7.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.4
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
networkx==3.1
nodeenv==1.8.0
oauthlib==3.2.2
objgraph==3.6.1
omegaconf==2.3.0
onnx==1.13.0
onnx-simplifier==0.4.35
onnxruntime==1.13.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip-tools==7.3.0
pipenv==2023.12.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pre-commit==3.5.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyaml==23.12.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==2.4.5
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.10.1
# seaborn==0.13.1
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stringcase==1.2.0
# super-gradients==3.1.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tomli==2.0.1
torchmetrics==0.8.0
tqdm==4.66.1
treelib==1.6.1
typing_extensions==4.9.0
tzdata==2023.4
# ultralytics @ git+https://github.com/mikel-brostrom/ultralytics.git@6d5a653254cd578ce0ab4e9270bdbecabee70c52
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
yacs==0.1.8
zipp==3.17.0
mysql-connector-python==8.3.0
gevent==24.2.1
gunicorn==21.2.0
Flask-HTTPAuth==4.8.0
itsdangerous==2.1.2
redis==5.0.3
sahi==0.11.18
flask_socketio==5.3.6
gevent-websocket==0.10.1
boxmot==10.0.63
kafka-python==2.2.15
numpy==1.24.4